pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "asyncssh>=2.21.0",
    "boto3>=1.38.30",
    "discord-py>=2.5.2",
    "flask>=3.1.1",
    "motor>=3.7.1",
    "py-cord==2.6.1",
    "pymongo>=4.13.0",
    "python-dotenv>=1.1.0",
]
